FishingScene.jsのリファクタリング gameファイルを作成し、処理ごとにファイル分けを行った。
